feat: add station coordinates fetching and routing matrix integration
- Introduced StationCoord struct and FetchAllStationCoords method to retrieve all station coordinates from the database.
- Added QueryStationCoords method to return a map of station IDs to their coordinates for fallback use.
- Updated PollBikes job to utilize station coordinates when bike location data is missing.
- Changed CurrentRangeMeters in Bike struct to float64 to comply with the specification.
- Enhanced trip detection logic in the Detector to handle bike ID rotations and utilize a routing matrix for travel time estimation.
- Implemented a new routing Matrix struct to fetch and store inter-station distances from OSRM.
- Added TravelWindow and RoadDistance methods to the Matrix for calculating travel durations and distances.
- Updated go.sum files to include new dependencies for the routing functionality.

+++ poller/internal/gbfs/models.go
@@ -85,7 +85,7 @@ type Bike struct {
 	IsDisabled         bool    `json:"is_disabled"`
 	VehicleTypeID      string  `json:"vehicle_type_id"`
 	LastReported       int64   `json:"last_reported"`
-	CurrentRangeMeters int     `json:"current_range_meters"`
+	CurrentRangeMeters float64 `json:"current_range_meters"` // spec: non-negative float
 	StationID          string  `json:"station_id"`
 	PricingPlanID      string  `json:"pricing_plan_id"`
 }
